refactor(backup/s3): add util class `FileSetManager`
